+++ include/sdf/Cylinder.hh
@@ -17,6 +17,7 @@
 #ifndef SDF_CYLINDER_HH_
 #define SDF_CYLINDER_HH_
 
+#include <ignition/math/Cylinder.hh>
 #include <sdf/Error.hh>
 #include <sdf/Element.hh>
 #include <sdf/sdf_config.h>
@@ -87,6 +88,14 @@ namespace sdf
     /// \return SDF element pointer. The value will be nullptr if Load has
     /// not been called.
     public: sdf::ElementPtr Element() const;
+
+    /// \brief Get the Ignition Math representation of this Cylinder.
+    /// \return A const reference to an ignition::math::Sphered object.
+    public: const ignition::math::Cylinderd &Shape() const;
+
+    /// \brief Get a mutable Ignition Math representation of this Cylinder.
+    /// \return A reference to an ignition::math::Cylinderd object.
+    public: ignition::math::Cylinderd &Shape();
 
     /// \brief Private data pointer.
     private: CylinderPrivate *dataPtr;