+++ src/Box_TEST.cc
@@ -41,6 +41,9 @@ TEST(DOMBox, MoveConstructor)
 
   sdf::Box box2(std::move(box));
   EXPECT_EQ(size, box2.Size());
+
+  EXPECT_DOUBLE_EQ(1 * 2 * 3, box2.Shape().Volume());
+  EXPECT_EQ(size, box2.Shape().Size());
 }
 
 /////////////////////////////////////////////////
@@ -103,7 +106,6 @@ TEST(DOMBox, CopyAssignmentAfterMove)
   EXPECT_EQ(size1, box2.Size());
 }
 
-
 /////////////////////////////////////////////////
 TEST(DOMBox, Load)
 {
@@ -130,4 +132,14 @@ TEST(DOMBox, Load)
   EXPECT_EQ(sdf::ErrorCode::ELEMENT_MISSING, errors[0].Code());
   EXPECT_NE(std::string::npos, errors[0].Message().find("missing a <size>"));
   EXPECT_NE(nullptr, box.Element());
+}
+
+/////////////////////////////////////////////////
+TEST(DOMBox, Shape)
+{
+  sdf::Box box;
+  EXPECT_EQ(ignition::math::Vector3d::One, box.Size());
+
+  box.Shape().SetSize(ignition::math::Vector3d(1, 2, 3));
+  EXPECT_EQ(ignition::math::Vector3d(1, 2, 3), box.Size());
 }